Refine envs & orca_gym mudule import
