+++ envs/__init__.py
@@ -1,0 +1,5 @@
+from .orca_gym_env import OrcaGymBaseEnv, RewardType
+from .orca_gym_local_env import OrcaGymLocalEnv
+from .orca_gym_remote_env import OrcaGymRemoteEnv
+
+__all__ = ['OrcaGymBaseEnv', 'RewardType', 'OrcaGymLocalEnv', 'OrcaGymRemoteEnv']
