+++ envs/openloong/openloong_env.py
@@ -12,12 +12,12 @@ from openloong_dyn_ctrl import OpenLoongWBC, OrcaGym_Interface, ButtonState
 import numpy as np
 from gymnasium.core import ObsType
 from orca_gym.utils import rotations
-from envs.orca_gym_env import OrcaGymRemoteEnv
+from envs import OrcaGymRemoteEnv, OrcaGymLocalEnv
 from typing import Optional, Any, SupportsFloat
 from gymnasium import spaces
 from orca_gym.devices.keyboard import KeyboardClient, KeyboardInput
 
-class OpenLoongEnv(OrcaGymRemoteEnv):
+class OpenLoongEnv(OrcaGymLocalEnv):
     """
     Control the walking of the OpenLoong robot.
